requirements.txt
requests==2.31.0
python-dotenv==1.0.0
tabulate==0.9.0
pathlib==1.0.1
pytest==7.4.0
pytest-cov==4.1.0
